implemmenting cancel reservation + syncronize the costs with the DB

+++ src/application/MainController.java
@@ -292,7 +292,7 @@ public class MainController {
 		}
 		LogInController logInController = loader.getController();
 		logInController.setWelcome("Welcome to CPS");
-		logInController.setTopOfLogInView(_username, Long.toString(_currentUser.getBalance()));
+		logInController.setTopOfLogInView(_username, Double.toString(_currentUser.getBalance()));
 		
 		if(_currentUser.getType().equals("b")){
 			Button businessButton = new Button("Business Routinely Subscription");
